Minor stuff fixed. Threads still missing.
Everything missing in the last commit is still missing now.

Basically we want a thread for each process, referencing random memory
addresses at a random pace. This kind of already exists in the
'development' branch, though bringing things from there require a lot
of adaptation (e.g. the page table functions). We may "merge" the thread
implementation, or just do it the dumb way (generate sequential random accesses on the main loop).

Also, the report is missing. And maybe slides.